A single uniform connection instead of separate connections in the every class
Connecting in every data class is a sure bad practice, and anyone who would try to extend the previously used approach to other classes will end up with tens of simultaneous connections to the database from the same PHP instance. 

So a database connection should be instantiated only once and then passed to all objects as a constructor parameter.

Given this lesson is linked on the PHPTheRightWay, it shouldn't promote a bad practice, even for sake of simplification.